Move state to ChangePointStates
